Remove broken status flag from persona state request
Steam's handling of the Status flag in ClientRequestFriendData appears to be broken, as it always returns persona_state: 0 (Offline).

Since Steam echoes ClientPersonaState messages to all clients logged in as the same account, this results in bad status data being sent to all clients. This can be verified by logging into the official Steam client, then logging in through steam-user, which causes all friends to appear to go offline.

Removing the status flag fixes this issue, and does not affect the availability of status data, as Steam automatically sends (correct) statuses for online friends.